+++ components/engine_traits/Cargo.toml
@@ -6,6 +6,7 @@ publish = false
 
 [features]
 default = ["protobuf-codec"]
+failpoints = ["fail/failpoints"]
 protobuf-codec = [
   "error_code/protobuf-codec",
   "raft/protobuf-codec",
@@ -35,6 +36,7 @@ slog = { version = "2.3", features = ["max_level_trace", "release_max_level_debu
 slog-global = { version = "0.1", git = "https://github.com/breeswish/slog-global.git", rev = "d592f88e4dbba5eb439998463054f1a44fbf17b9" }
 kvproto = { git = "https://github.com/pingcap/kvproto.git", default-features = false }
 raft = { version = "0.6.0-alpha", default-features = false }
+fail = "0.4"
 
 [dev-dependencies]
 toml = "0.5"
