+++ src/server/ttl/mod.rs
@@ -5,3 +5,6 @@ mod ttl_compaction_filter;
 
 pub use ttl_checker::{TTLChecker, Task as TTLCheckerTask};
 pub use ttl_compaction_filter::TTLCompactionFilterFactory;
+
+#[cfg(feature = "failpoints")]
+pub use ttl_checker::check_ttl_and_compact_files;
