+++ src/storage/txn/commands/atomic_store.rs
@@ -2,12 +2,12 @@
 
 use crate::storage::kv::{Modify, WriteData};
 use crate::storage::lock_manager::LockManager;
-use crate::storage::raw::ttl::convert_to_expire_ts;
 use crate::storage::txn::commands::{
     Command, CommandExt, ResponsePolicy, TypedCommand, WriteCommand, WriteContext, WriteResult,
 };
 use crate::storage::txn::Result;
 use crate::storage::{ProcessResult, Snapshot};
+use engine_traits::util::convert_to_expire_ts;
 use engine_traits::CfName;
 use txn_types::Mutation;
