+++ tests/failpoints/cases/mod.rs
@@ -20,3 +20,4 @@ mod test_stale_read;
 mod test_storage;
 mod test_transaction;
 mod test_transfer_leader;
+mod test_ttl;
